requirements.txt
certifi==2025.8.3
elastic-transport==8.11.0
elasticsearch==8.11.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions>=4.8.0
urllib3==2.5.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
bertopic>=0.15.0
scikit-learn>=1.0.0
transformers>=4.0.0
torch>=1.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
sentence-transformers>=2.2.0
hdbscan>=0.8.0
umap-learn>=0.5.0
